+++ src/models/EmailTestMode.ts
@@ -1,0 +1,4 @@
+export interface EmailTestMode {
+    emailTestMode: boolean;
+    emailTestAddress?: string
+}